LayoutTests/imported/w3c/IndexedDB-private-browsing/idbcursor_continue_index3-expected.txt


PASS IDBCursor.continue() - index - attempt to iterate to the previous record when the direction is set for the next record
